pyproject.toml
[tool.poetry]
name = "imobi-ai-crm"
version = "0.2.0"
description = "Sistema CRM Imobiliário com IA"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
sqlalchemy = "^2.0.42"
psycopg2-binary = "^2.9.10"
alembic = "^1.14.1"
crewai = "^0.150.0"
langchain = "^0.3.27"
langchain-openai = "^0.3.28"
langchain-core = "^0.3.72"
langchain-community = "^0.3.27"
openai = "^1.97.1"
tiktoken = "^0.9.0"
pandas = "^2.2.3"
numpy = "^1.26.4"
lxml = "^5.3.0"
httpx = "^0.28.1"
requests = "^2.32.4"
aiohttp = "^3.12.15"
twilio = "^9.7.0"
celery = "^5.4.0"
redis = "^5.2.1"
python-dotenv = "^1.1.1"
python-multipart = "^0.0.20"
jinja2 = "^3.1.6"
python-dateutil = "^2.9.0"
loguru = "^0.7.3"
bcrypt = "^4.3.0"
cryptography = "^45.0.5"
openpyxl = "^3.1.5"
pillow = "^11.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
